smart_city/requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
sqlparse==0.5.1
tzdata==2024.1
django-filter
python-dateutil==1.4